streamline SwapQuoter and remove instant-only functions
